# [multi-property] performance improvement