- file.c
```c
int main()
{
    for(int i =0 ; i<10;i++)
        for(int j =0 ; j<10; j++)
            for(int k =0 ; k<10; k++)
                for(int l =0 ; l<10; l++)
                    assert(i+j+k>l);
}
```
- command: `./esbmc file.c --unwind 12 --multi-property`
- result (debug version)
```sh
# The verification result is the same

# Before
./esbmc file.c --unwind 12 --multi-property  50.69s user 0.55s system 99% cpu 51.267 total
# After
./esbmc file.c --unwind 12 --multi-property  13.25s user 0.56s system 99% cpu 13.844 total
```
- The output is also simplified and beautified, no more annoying "Found verified claim. Skipping..."