@Anthonysdu 2024-01-20
> @Anthonysdu: could you please run this PR over the RMM project?

The master branch cannot process the rmm code since PR #1602 :
```
Converting
Generating GOTO Program
GOTO program creation time: 3.414s
Computing Value-Set Analysis (VSA)
Segmentation fault (core dumped)
```
PR #1601 :

```
Converting
Generating GOTO Program
GOTO program creation time: 3.419s
Computing Value-Set Analysis (VSA)
GOTO program processing time: 7.505s
Starting Bounded Model Checking
WARNING: no body for function __builtin_clzll
WARNING: no body for function __builtin_clzll
Unwinding loop 53 iteration 1   file 
```