@rafaelsamenezes 2024-01-20
@Anthonysdu could you check if this PR solves the issues: https://github.com/esbmc/esbmc/pull/1607?

Also, it would be great if you could generate a test case that is causing the performance degradation