@rafaelsamenezes 2024-01-20
Great addition, I would suggest moving the logic to here: https://github.com/esbmc/esbmc/blob/d966de633efccd2074aa53b4904631ad66a747bc/src/esbmc/bmc.cpp#L735

Mainly because:

1. I don't think you need to rely on the claim slicer to get the assert. You can get it directly from VCC.
2. The list of jobs should be filtered before, specially if we want for this to eventually work in parallel.