@lucasccordeiro 2024-01-21
@Anthonysdu: can you test this PR performance without the interval analysis enabled?