@rafaelsamenezes 2024-01-21
> > @Anthonysdu could you check if this PR solves the issues: #1607?
> > 
> > Also, it would be great if you could generate a test case that is causing the performance degradation
> 
> ```
> Converting
> Generating GOTO Program
> GOTO program creation time: 2.989s
> Segmentation fault (core dumped)
> ```
> It still happens, but can get rid of it if disable `--interval-analysis`

I guess we should move to a new issue. Could you run with a debugger and return the back trace