@Anthonysdu 2024-01-22
```
before:
			claims		time(s)
RMI_REC_DESTROY		113		2579.9
RMI_GRANULE_DELEGATE	54		357.7
RMI_GRANULE_UNDELEGATE	45 		289.1
RMI_REALM_ACTIVATE	53		530.4
RMI_REALM_DESTROY	114		1659.8
RMI_REC_AUX_COUNT	48 		507
RMI_FEATURES		21 		276.8
RMI_DATA_DESTROY	82		>5000

This PR:
			claims		time(s)
RMI_REC_DESTROY		84		1477.3
RMI_GRANULE_DELEGATE	54		394.8
RMI_GRANULE_UNDELEGATE	45 		341.2
RMI_REALM_ACTIVATE	53		503.5
RMI_REALM_DESTROY	72		792.2
RMI_REC_AUX_COUNT	48 		441.9
RMI_FEATURES		21 		210.4
RMI_DATA_DESTROY	117		>5000
```

claims in two testcases `rec_destory` and `realm_destory` are reduced.