@lucasccordeiro 2024-01-22
Many thanks, @Anthonysdu. What about the number of safe and unsafe cases reported by esbmc? Did they change?