@fbrausse 2024-01-22
> claims in two testcases `rec_destory` and `realm_destory` are reduced.

But RMI_DATA_DESTROY did go up. @ChenfengWei0 Do you know why they might go up with this PR?