@Anthonysdu 2024-01-22
> > claims in two testcases `rec_destory` and `realm_destory` are reduced.
> 
> But RMI_DATA_DESTROY did go up. @ChenfengWei0 Do you know why they might go up with this PR?

This case, the previous run did not stop in 7 hours (if you could remember), and the new run was killed by system in 2 hours because out of 16GB memory, so actually both run did not finish.

We also have the run realm_destory being killed twice. And this PR reports more bugs in the two cases and I need to confirm them.